Don't run TestMetadataChanged test on Windows
